Completed implementation of the second activity screen so that users can see the stats going down as time goes by.
The big challenge here was learning how to pass an object between activities.
This was solved by having the jake class implement Parcelable, using
auto-complete to add the bare -bones methods.

After this accessing the data within my fresh_dog instance was easy, and
the results are now displayed on screen.