I am continuing through an Udemy course on creating pixel art so that I can draw my own animations.
I got tired of the purple screen so created a surrealist room for the dog
to live in for now. As I continue to get better at photoshop I will
update.

Hoping to be through this course in a month or so (with the little free
time I have) so should have some dog / creature animations very soon.