In order to make some progress I am going to use my shodily drawn images of the dog.
We now have an initialised little dog sitting waiting for some interaction.
If you press the play button it will start a tester animation to make
sure it at least animates. If you press either "feed" or "water" then the
dog will sit back down.

Next push will be to have the dog walking left and right, and facing in
the right direction.